+++ thonny/plugins/backend_config_page.py
@@ -81,7 +81,7 @@ class BackendConfigurationPage(ConfigurationPage):
         self._combo_variable = create_string_var(current_backend_desc)
 
         label = ttk.Label(
-            self, text=tr("Which interpreter or device should Thonny use for running your code?")
+            self, text=tr("Which kind of interpreter should Thonny use for running your code?")
         )
         label.grid(row=0, column=0, columnspan=2, sticky=tk.W)